Refactor new feature
