+++ x-pack/plugin/core/src/main/resources/META-INF/services/org.elasticsearch.features.FeatureSpecification
@@ -5,4 +5,4 @@
 # 2.0.
 #
 
-org.elasticsearch.xpack.core.datatiers.DataTiersFeatures
+org.elasticsearch.xpack.core.XPackFeatures
